requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.2
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
lxml==5.2.1
oauth2client==4.1.3
oauthlib==3.2.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
simplegmail==4.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
